moved redis file to utils
