+++ src/utils/redis/index.ts
@@ -1,0 +1,1 @@
+export { redis } from './redis';
